#24 refactor: BaseFragment 적용 및 DateTimeUtil app 모듈로 이동
