+++ data/build.gradle.kts
@@ -47,6 +47,7 @@ android {
 }
 
 dependencies {
+    implementation(project(":app"))
     implementation(project(":domain"))
 
     implementation(libs.androidx.appcompat)
