+++ presentation/src/main/java/com/nextroom/nextroom/presentation/ui/adminmain/AdminMainState.kt
@@ -3,7 +3,7 @@ package com.nextroom.nextroom.presentation.ui.adminmain
 import com.nextroom.nextroom.domain.model.UserSubscribeStatus
 import com.nextroom.nextroom.presentation.extension.calculateDday
 import com.nextroom.nextroom.presentation.model.ThemeInfoPresentation
-import com.nextroom.nextroom.presentation.util.DateTimeUtil
+import com.nextroom.nextroom.util.DateTimeUtil
 
 data class AdminMainState(
     val loading: Boolean = false,
@@ -12,6 +12,7 @@ data class AdminMainState(
     val themes: List<ThemeInfoPresentation> = emptyList(),
 ) {
     private val dateTimeUtil = DateTimeUtil()
+
     fun calculateDday(): Int {
         return dateTimeUtil.stringToDate(userSubscribeStatus.expiredDate, "yyyy.MM.dd")?.calculateDday() ?: -1
     }
