--- presentation/src/main/java/com/nextroom/nextroom/presentation/ui/mypage/MypageEvent.kt
@@ -1,5 +1,0 @@
-package com.nextroom.nextroom.presentation.ui.mypage
-
-sealed interface MypageEvent {
-    data object Logout : MypageEvent
-}
